Lagt til iverksetting av barnetilsyn

+++ src/main/kotlin/no/nav/familie/ef/sak/vedtak/domain/Vedtak.kt
@@ -52,7 +52,7 @@ data class Barnetilsynperiode(val datoFra: LocalDate,
 
 data class PeriodeMedBeløp(val datoFra: LocalDate,
                            val datoTil: LocalDate,
-                           val beløp: BigDecimal) {
+                           val beløp: Int) {
 
     fun tilDto() = PeriodeMedBeløpDto(årMånedFra = YearMonth.from(datoFra),
                                       årMånedTil = YearMonth.from(datoTil),
